Update getMessageType function and messagePreview
