+++ script/utils/message.js
@@ -1,7 +1,7 @@
-import {wrapAsObjWithKey, formatDate, tgDate} from '../helpers.js';
+import {formatDate, tgDate, wrapAsObjWithKey} from '../helpers.js';
 import {userIdToPeerId} from '../../state/users/utils.js';
 
-const {isObjectOf, construct} = zagram;
+const {isObjectOf, construct, CONSTRUCTOR_KEY} = zagram;
 
 const getMedia = R.prop('media');
 
@@ -161,3 +161,121 @@ export const getMessageDateStr = R.pipe(
   getMessageDate,
   formatDate
 );
+
+
+export const getStickerEmojiFromMessage = R.pipe(
+  R.pathOr([], ['media', 'document', 'attributes']),
+  R.filter(isObjectOf('documentAttributeSticker')),
+  R.nth(0),
+  R.propOr('', 'alt')
+);
+
+
+export const isAudioDocument = R.pipe(
+  R.prop('attributes'),
+  R.filter(isObjectOf('documentAttributeAudio')),
+  R.length,
+  R.equals(1)
+);
+
+
+/**
+ * Returns true if document is sticker
+ */
+export const isVideoDocument = R.pipe(
+  R.prop('attributes'),
+  R.filter(isObjectOf('documentAttributeVideo')),
+  R.length,
+  R.equals(1)
+);
+
+
+/**
+ * Returns true if document is sticker
+ */
+export const isStickerDocument = R.pipe(
+  R.prop('attributes'),
+  R.filter(isObjectOf('documentAttributeSticker')),
+  R.length,
+  R.equals(1)
+);
+
+
+/**
+ * Returns type of message from document
+ * @param {*} - Document
+ */
+export const getTypeFromDocument = R.cond([
+  [isVideoDocument, R.always('messageVideo')],
+  [isAudioDocument, R.always('messageAudio')],
+  [isStickerDocument, R.always('messageSticker')],
+  [R.T, R.always('messageMediaDocument')]
+]);
+
+
+/**
+ * Returns 'messageText' if message without media else returns type by
+ * message media
+ * @param media
+ */
+export const getTypeFromMedia = R.cond([
+  [isObjectOf('messageMediaEmpty'), R.always('messageText')],
+  [isObjectOf('messageMediaDocument'), R.pipe(R.prop('document'), getTypeFromDocument)],
+  [R.T, R.prop(CONSTRUCTOR_KEY)]
+]);
+
+
+/**
+ * Returns 'messageText' if message without media else returns type by
+ * message media
+ * @param message
+ */
+export const getCommonMessageType = R.cond([
+  [R.has('media'), R.pipe(R.prop('media'), getTypeFromMedia)],
+  [R.T, R.always('messageText')]
+]);
+
+
+/**
+ * Returns service message type by action constructor:
+ * https://core.telegram.org/type/MessageAction
+ * @param message
+ */
+export const getServiceMessageType = R.path(['action', CONSTRUCTOR_KEY]);
+
+
+/**
+ * Returns types of message for function
+ * @param message
+ */
+export const getMessageType = R.cond([
+  [isObjectOf('message'), getCommonMessageType],
+  [isObjectOf('messageService'), getServiceMessageType]
+]);
+
+
+/**
+ * @param {*} message - telegram message or service message object
+ * @return {string} - text preview for last message
+ */
+export function previewMessage(message) {
+  const type = getMessageType(message);
+  const handlers = {
+    messageText: R.prop('message'),
+    messageMediaEmpty: R.prop('message'),
+    messageMediaWebPage: R.prop('message'),
+    messageMediaPoll: () => '📊 Poll',
+    messageMediaPhoto: m => '🖼 ' + (m.message || 'Photo'),
+    messageVideo:     m => '🎥 ' + (getFilename(m) || 'Video'),
+    messageAudio:     m => '🎵 ' + (getFilename(m) || 'Audio'),
+    messageSticker: m => 'Sticker ' + getStickerEmojiFromMessage(m),
+    messageMediaDocument: () => `Document`,
+    // messageAnimation: c => 'GIF',
+    messageChatAddMembers: () => 'добавление в чат',
+    messageActionContactSignUp: () => 'теперь в телеграм'
+  };
+  const text = typeof handlers[type] === 'function'
+    ? handlers[type](message)
+    : 'неподдерживаемое сообщение (' + type + ')';
+  return text.split(/\n/)[0];
+}
