+++ state/dialogs/helpers.js
@@ -4,7 +4,7 @@
 import {peerIdToPeer} from '../utils.js';
 import {wrapAsObjWithKey} from '../../script/helpers.js';
 
-const {isObjectOf, construct, CONSTRUCTOR_KEY} = zagram;
+const {isObjectOf, construct} = zagram;
 
 export const getLastMessage = R.pipe(
   R.of,
@@ -174,103 +174,6 @@ export const getDialogTitle = R.pipe(
   getTitle
 );
 
-
-/**
- * Returns true if document is sticker
- */
-export const isVideoDocument = R.pipe(
-  R.prop('attributes'),
-  R.filter(isObjectOf('documentAttributeVideo')),
-  R.length,
-  R.equals(1)
-);
-
-
-/**
- * Returns true if document is sticker
- */
-export const isStickerDocument = R.pipe(
-  R.prop('attributes'),
-  R.filter(isObjectOf('documentAttributeSticker')),
-  R.length,
-  R.equals(1)
-);
-
-
-/**
- * Returns type of message from document
- * @param {*} - Document
- */
-export const getTypeFromDocument = R.cond([
-  [isVideoDocument, R.always('messageVideo')],
-  [isStickerDocument, R.always('messageSticker')],
-  [R.T, R.always('messageMediaDocument')]
-])
-
-
-/**
- * Returns 'messageText' if message without media else returns type by
- * message media
- * @param media
- */
-export const getTypeFromMedia = R.cond([
-  [isObjectOf('messageMediaEmpty'), R.always('messageText')],
-  [isObjectOf('messageMediaDocument'), R.pipe(R.prop('document'), getTypeFromDocument)],
-  [R.T, R.prop(CONSTRUCTOR_KEY)]
-]);
-
-
-/**
- * Returns 'messageText' if message without media else returns type by
- * message media
- * @param message
- */
-export const getCommonMessageType = R.cond([
-  [R.has('media'), R.pipe(R.prop('media'), getTypeFromMedia)],
-  [R.T, R.always('messageText')]
-]);
-
-
-/**
- * Returns service message type by action constructor:
- * https://core.telegram.org/type/MessageAction
- * @param message
- */
-export const getServiceMessageType = R.path(['action', CONSTRUCTOR_KEY]);
-
-/**
- * Returns types of message for function
- * @param message
- */
-export const getMessageType = R.cond([
-  [isObjectOf('message'), getCommonMessageType],
-  [isObjectOf('messageService'), getServiceMessageType]
-]);
-
-/**
- * @param {*} message - telegram message or service message object
- * @return {string} - text preview for last message
- */
-export function previewMessage(message) {
-  const type = getMessageType(message);
-  const handlers = {
-    messageText: m => m.message,
-    messageMediaEmpty: m => m.message,
-    messageMediaPoll: () => '📊 Poll',
-    messageMediaPhoto:     m => '🖼 ' + (m.message || 'Photo'),
-    // messageVideo:     c => '🎥 ' + (c.caption && c.caption.text || 'Video'),
-    // messageAudio:     c => '🎵 ' + c.audio.title || 'Audio',
-    // messageSticker:   c =>  'Sticker ' + c.sticker.emoji,
-    messageMediaDocument:  () => `Document`,
-    // messageAnimation: c => 'GIF',
-    messageChatAddMembers:    () => 'добавление в чат',
-    messageActionContactSignUp: () => 'теперь в телеграм'
-  };
-  const text = typeof handlers[type] === 'function'
-    ? handlers[type](message)
-    : 'неподдерживаемое сообщение (' + type + ')';
-  return text.split(/\n/)[0];
-}
 
 export const getTitleFromPeerInfo = R.pipe(
   R.prop('base'),
